M116 public

+++ scripts/build_Linux.sh
@@ -14,13 +14,16 @@ if [[ $(uname -m) == "aarch64" ]]; then
 fi
 
 # Install system dependencies
-yum install -y \
-    fontconfig-devel \
-    mesa-libGL-devel \
-    xorg-x11-server-Xvfb \
-    mesa-dri-drivers && \
-    yum clean all && \
-    rm -rf /var/cache/yum
+if [[ $EUID -eq 0 ]]; then
+    yum install -y \
+        python3 \
+        fontconfig-devel \
+        mesa-libGL-devel \
+        xorg-x11-server-Xvfb \
+        mesa-dri-drivers && \
+        yum clean all && \
+        rm -rf /var/cache/yum
+fi
 
 if [[ $(uname -m) == "aarch64" ]] && [[ $CI_SKIP_BUILD == "true" ]]; then
     # gn and skia already built in a previous job
@@ -35,28 +38,28 @@ export CFLAGS="-Wno-deprecated-copy"
 export LDFLAGS="-lrt"
 git clone https://gn.googlesource.com/gn && \
     cd gn && \
-    git checkout 981f46c64d1456d2083b1a2fa1367e753e0cdc1b && \
+    git checkout fe330c0ae1ec29db30b6f830e50771a335e071fb && \
     python build/gen.py && \
     ninja -C out gn && \
     cd ..
 
 # Build skia
 cd skia && \
-    patch -p1 < ../patch/git-sync-deps.patch && \
-    python tools/git-sync-deps && \
-    patch -p1 < ../patch/make_data_assembly.patch && \
-    patch -p1 < ../patch/libjpeg-arm.patch && \
+    patch -p1 < ../patch/skia-m116-minimize-download.patch && \
+    patch -p1 < ../patch/skia-m116-colrv1-freetype.diff && \
+    python3 tools/git-sync-deps && \
     cp -f ../gn/out/gn bin/gn && \
     bin/gn gen out/Release --args="
 is_official_build=true
-skia_enable_tools=true
+skia_enable_svg=true
 skia_use_system_libjpeg_turbo=false
 skia_use_system_libwebp=false
 skia_use_system_libpng=false
 skia_use_system_icu=false
 skia_use_system_harfbuzz=false
+skia_use_system_freetype2=false
 extra_cflags_cc=[\"-frtti\"]
 extra_ldflags=[\"-lrt\"]
 " && \
-    ninja -C out/Release skia skia.h experimental_svg_model && \
+    ninja -C out/Release && \
     cd ..
